(tests) Remove convergence test for Akaike bins
The condition is not true. The correction does not vanish for large
bin counts which makes it a bad test. Additionally, we don't use the
Small sample correction in our tests so it doesn't matter and I don't
want to deal with this failing test case anymore.

Maybe I should also delete the small sample test, idk. But honestly,
whatever.